+++ src/plugin.hpp
@@ -10,6 +10,7 @@ using namespace rack;
 extern Plugin* pluginInstance;
 
 // Declare each Model, defined in each module source file
+extern Model* modelAD_ASR;
 extern Model* modelAuxClone;
 extern Model* modelBayInput;
 extern Model* modelBayNorm;